+++ Model/dettagli.h
@@ -11,7 +11,7 @@ class Dettagli
 private:
     Model *model;
     double raggio;
-    int temp;
+    double temp;
     double dens;
     unsigned int eta;
     double velRot;
@@ -29,9 +29,9 @@ public:
     Dettagli(Model *, const OggettoCeleste * const);
 
     double getRaggio() const;
-    int getTemp() const;
+    double getTemp() const;
     double getDens() const;
-    unsigned long int getEta() const;
+    unsigned int getEta() const;
     double getVelRot() const;
     double getSemiA() const;
     double getAz() const;
