+++ Model/stella.h
@@ -8,7 +8,7 @@ private:
     double magnitudineAss; //cioè magnitudine a 10 parsec di distanza, da -20 a 10
     double magnitudineApp; // cioè magnitudine dalla terra, da -26.74(sole) a 10
 public:
-    Stella(long double, int, double, unsigned int, double, double, double);
+    Stella(long double, double, double, unsigned int, double, double, double);
     Stella(const CorpoCeleste&, double, double);
 
     double getMAss() const;
